Fix checkstyle issues
Split AbstractDriverFactoryTest and WebDriverFactoryTest tests

Move firefox capabilities creating methods from AbstractDriverFactory
