+++ src/test/java/com/codeborne/selenide/webdriver/WebDriverFactoryTest.java
@@ -6,22 +6,13 @@ import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.Dimension;
 import org.openqa.selenium.Point;
-import org.openqa.selenium.Proxy;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.firefox.FirefoxProfile;
 
-import java.io.IOException;
-
-import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.*;
 
 public class WebDriverFactoryTest {
   WebDriverFactory factory = spy(new WebDriverFactory());
   WebDriver webdriver = mock(WebDriver.class, RETURNS_DEEP_STUBS);
-  Proxy proxy = mock(Proxy.class);
 
   @Before
   public void setUp() {
@@ -52,73 +43,6 @@ public class WebDriverFactoryTest {
     factory.adjustBrowserSize(webdriver);
 
     verify(webdriver.manage().window()).maximize();
-  }
-
-  @Test
-  public void transfersStringCapabilitiesFromSystemPropsToDriver() {
-    System.setProperty("capabilities.some.cap", "abcd");
-
-    final AbstractDriverFactory driverFactory = new DefaultDriverFactory();
-
-    assertThat(driverFactory.createCommonCapabilities(proxy).getCapability("some.cap"), is("abcd"));
-  }
-
-  @Test
-  public void transfersBooleanCapabilitiesFromSystemPropsToDriver() {
-    System.setProperty("capabilities.some.cap", "true");
-
-    final AbstractDriverFactory driverFactory = new DefaultDriverFactory();
-
-    assertThat(driverFactory.createCommonCapabilities(proxy).getCapability("some.cap"), is(true));
-  }
-
-  @Test
-  public void transfersIntegerCapabilitiesFromSystemPropsToDriver() {
-    System.setProperty("capabilities.some.cap", "25");
-
-    final AbstractDriverFactory driverFactory = new DefaultDriverFactory();
-
-    assertThat(driverFactory.createCommonCapabilities(proxy).getCapability("some.cap"), is(25));
-  }
-
-  @Test
-  public void transferIntegerFirefoxProfilePreferencesFromSystemPropsToDriver() {
-    System.setProperty("firefoxprofile.some.cap", "25");
-
-    final AbstractDriverFactory driverFactory = new DefaultDriverFactory();
-
-    FirefoxProfile profile = (FirefoxProfile) driverFactory.createFirefoxCapabilities(proxy).getCapability(FirefoxDriver.PROFILE);
-    assertThat(profile.getIntegerPreference("some.cap", 0), is(25));
-
-  }
-
-  @Test
-  public void transferBooleanFirefoxProfilePreferencesFromSystemPropsToDriver() {
-    System.setProperty("firefoxprofile.some.cap", "false");
-
-    final AbstractDriverFactory driverFactory = new DefaultDriverFactory();
-
-    FirefoxProfile profile = (FirefoxProfile) driverFactory.createFirefoxCapabilities(proxy).getCapability(FirefoxDriver.PROFILE);
-    assertThat(profile.getBooleanPreference("some.cap", true), is(false));
-  }
-
-  @Test
-  public void transferStringFirefoxProfilePreferencesFromSystemPropsToDriver() {
-    System.setProperty("firefoxprofile.some.cap", "abdd");
-
-    final AbstractDriverFactory driverFactory = new DefaultDriverFactory();
-
-    FirefoxProfile profile = (FirefoxProfile) driverFactory.createFirefoxCapabilities(proxy).getCapability(FirefoxDriver.PROFILE);
-    assertThat(profile.getStringPreference("some.cap", "sjlj"), is("abdd"));
-  }
-
-  @Test
-  public void transferChromeOptionArgumentsFromSystemPropsToDriver() throws IOException {
-    System.setProperty("chromeoptions.args", "abdd,--abcd,xcvcd=123");
-    String arrayOfArguments = new ChromeDriverFactory().createChromeOptions().toJson().getAsJsonObject().getAsJsonArray("args").toString();
-    assertThat(arrayOfArguments, containsString("abdd"));
-    assertThat(arrayOfArguments, containsString("--abcd"));
-    assertThat(arrayOfArguments, containsString("xcvcd=123"));
   }
 
   @Test
